Add latest post excerpt to front page
Used Jekyll's liquid tags to find the latest blog post and insert it (with some metadata) at the end of the site's front page. Since this is meant to only pull an excerpt of the post, only the first paragraph of the post will be shown unless an excerpt tag is explicitly placed elsewhere in the post.