Refactor stylesheets
Separated out styles that are only needed for the front page into frontpage.css. It's not worth loading all of these extra styles for the majority of the pages on the site that don't need them.

Also changed the wrapper on blog posts to .post-content to create consistency between the front page and the blog post layout.