Add black tint to hero image
The white hero text was clashing with the white bike in the hero image,
so adding a semi-transparent black tint to overlay on the hero image
gives a little contrast and makes the text more readable. Bonus: the
gradient that makes this work can have colors added to it if I decide
to get a little fancier later.