Fix footer not sticking to bottom of view
I'm using a giant flexbox over the whole body of the site to achieve
the vertical spread needed for the layout. The main contents of a
page will be allowed to flex-grow to fill the whole height of the view
and will push the footer to the bottom as a result.

Side effect: the whole site can be more easily rearranged for larger
screens if I decide to add sidebars, etc.